Moved resource controllers into dedicated folder
