+++ app/Http/Controllers/Resources/DashboardController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Resources;
 
 use Inertia\Inertia;
