+++ app/Http/Controllers/Resources/PublicPagesController.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Resources;
 
 use App\Helpers\SEO;
+use App\Http\Controllers\Controller;
 use App\Repositories\ArticleRepository;
 use App\Repositories\ConfigurationRepository;
 use App\Repositories\PageRepository;
