+++ app/Http/Controllers/Resources/SitemapController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Resources;
 
 use App\Repositories\ArticleRepository;
 use App\Repositories\CategoryRepository;
