+++ app/Http/Controllers/Resources/UserController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Resources;
 
 use App\Helpers\PaginatedCollection;
 use App\Http\Controllers\Base\ResourceController;
